Added Level Editor
Level Editor allows the user to create, save and load maps. Press Q, W,
or E to change the type of block to be placed. Hold S and click on top
of any cube to add a cube to the map. Press save to save to 'test.map',
type 'test.map' into the textbox and press Load to load.